+++ packages/bridge-status-controller/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Support persisting and emitting Unified SwapBridge input currency mode on submitted/completed metrics ([#8909](https://github.com/MetaMask/core/pull/8909))
+- Support persisting and emitting Unified SwapBridge input currency mode on submitted/completed/failed metrics ([#8909](https://github.com/MetaMask/core/pull/8909))
 
 ### Changed